- Adding AX-B residual portion to DQMOM class - Fixing some things that give compiler errors (in KobayashiSarofimDevol, DQMOMEqnFactory, and DQMOMEqn classes)

+++ src/CCA/Components/Arches/TransportEqns/DQMOMEqn.h
@@ -180,7 +180,7 @@ public:
     return d_weight; };
 
   /** @brief Get the low clipping values */ 
-  inline int getLowClip(){
+  inline double getLowClip(){
     return d_lowClip; };
 
   /** @brief Sets this equation as a weight.
